+++ src/main/kotlin/br/com/tagliaferrodev/samplerest/repositories/NacionalidadeRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface NacionalidadeRepository : JpaRepository<Nacionalidade, Long>
+interface NacionalidadeRepository : JpaRepository<Nacionalidade, Int>
