+++ src/main/kotlin/br/com/tagliaferrodev/samplerest/repositories/UsuarioRepository.kt
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository
 import java.util.*
 
 @Repository
-interface UsuarioRepository : JpaRepository<Usuario, Long> {
+interface UsuarioRepository : JpaRepository<Usuario, Int> {
 
     fun findByEmailOrApelidoAndBloqueadoFalse(email: String, apelido: String): Optional<Usuario>
 }
